Break out bluetooth apis into api.py
Like #82291 this is not a functional change.
